fix: variable type

+++ src/Point.php
@@ -7,17 +7,17 @@ namespace Eliepse\WritingGrid;
 class Point
 {
 	/**
-	 * @var int
+	 * @var float
 	 */
 	public $x = 0;
 
 	/**
-	 * @var int
+	 * @var float
 	 */
 	public $y = 0;
 
 
-	public function __construct(int $x = 0, int $y = 0)
+	public function __construct(float $x = 0.0, float $y = 0.0)
 	{
 		$this->x = $x;
 		$this->y = $y;
